requirements.txt
Django==1.11
Padding==0.4
#Pillow
Unidecode==0.04.20
argparse==1.4.0
arrow==0.10.0
dj-database-url==0.4.2
dj-static==0.0.6
django-cors-headers==2.0.2
django-debug-toolbar==1.7
django-extensions
djangorestframework
markdown
markdown
django-filter
gunicorn==19.7.1
psycopg2==2.7.1
pystache==0.5.4
python-dateutil==2.6.0
pytz==2014.4
raven==5.32.0
simplejson==3.6.0
six==1.10.0
sqlparse==0.2.3
static==1.0.2
static3==0.7.0
validate-email==1.2
wsgiref==0.1.2
django-grappelli==2.8.1
Pygments
requests==2.13.0
urllib3==1.18
